Set up Cloudfront for oaklandcouncil.net
This adds Terraform resources for a TLS cert and accompanying Cloudfront
distribution for Councilmatic (doing business as OaklandCouncil.net).

Eventually it'd be swell to use the Terraform Namecheap provider to
create the DNS records to validate the TLS cert and point to
Cloudfront... but alas Namecheap has not approved our API request yet.